Factor tt::is_complex_of_fundamental into its own file
